backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "clerk-backend-api>=1.0.0",  # Changed from clerk-sdk-python
    "cryptography>=42.0.0",  # For API key encryption
    "fastapi>=0.125.0",
    "ipython>=9.8.0",
    "langchain-anthropic>=0.1.15",
    "langchain-core>=0.2.8",
    "langgraph>=0.2.0",  # Updated
    "manim>=0.19.1",
    "markdown>=3.10",
    "pdf2image>=1.17.0",
    "pillow>=10.0.0",
    "pypdf2>=3.0.0",
    "img2pdf>=0.5.0",
    "pydantic>=2.7.4",  # Updated to match langgraph requirements
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.21",
]
